% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{get_adaptive_penalties_internal_NWML_L2}
\alias{get_adaptive_penalties_internal_NWML_L2}
\title{Internal: compute adaptive penalties C1, C2, C3}
\usage{
get_adaptive_penalties_internal_NWML_L2(krtexas_model, A, int_ind_ml)
}
\description{
Internal: compute adaptive penalties C1, C2, C3
}
\keyword{internal}
