% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{get_gradients_wrtleaves_internal}
\alias{get_gradients_wrtleaves_internal}
\title{Internal: gradient wrt leaves for a single newx}
\usage{
get_gradients_wrtleaves_internal(krtexas_model, newx)
}
\description{
Internal: gradient wrt leaves for a single newx
}
\keyword{internal}
