% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{get_row_depths_from_root_internal}
\alias{get_row_depths_from_root_internal}
\title{Internal: compute depths of rows of A from root}
\usage{
get_row_depths_from_root_internal(A)
}
\description{
Internal: compute depths of rows of A from root
}
\keyword{internal}
