% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{get_siblings_m_internal}
\alias{get_siblings_m_internal}
\title{Internal: get siblings of a row in A}
\usage{
get_siblings_m_internal(A, m, depths)
}
\description{
Internal: get siblings of a row in A
}
\keyword{internal}
