% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{grad_fun}
\alias{grad_fun}
\title{Gaussian kernel gradient (internal, C++ wrapper)}
\usage{
grad_fun(u, w, Ax, X, Y, sigma, kappa, adaptive_weights)
}
\description{
Gaussian kernel gradient (internal, C++ wrapper)
}
\keyword{internal}
