% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{grad_fun_ep}
\alias{grad_fun_ep}
\title{Epanechnikov kernel gradient (internal)}
\usage{
grad_fun_ep(u, w, Ax, X, Y, sigma, kappa)
}
\description{
Epanechnikov kernel gradient (internal)
}
\keyword{internal}
