% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{krtexas_fit_internal}
\alias{krtexas_fit_internal}
\title{Internal KR TEXAS fitter for a given set of adaptive weights (with final fit)}
\usage{
krtexas_fit_internal(
  X,
  Y,
  A,
  kernel = "gaussian",
  nfolds = 4,
  lambda = NULL,
  min_lambda = 0,
  max_lambda = NULL,
  nlambda = 10,
  alpha = 0,
  eps = 0.01,
  gamma_threshold = 1e-06,
  use_num_grad = FALSE,
  silent = TRUE,
  parallel = TRUE,
  n_cores = NULL,
  warm_start = FALSE,
  adaptive_weights,
  gamma_init_strat = "smallest",
  developer_code = 0
)
}
\description{
Internal KR TEXAS fitter for a given set of adaptive weights (with final fit)
}
\keyword{internal}
