% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{krtexas_fixed_lambda}
\alias{krtexas_fixed_lambda}
\title{Internal solver for fixed lambda}
\usage{
krtexas_fixed_lambda(
  X = NULL,
  Y = NULL,
  A = NULL,
  kernel = "gaussian",
  lambda = 0,
  alpha = NULL,
  eps = 0.01,
  gamma_threshold,
  use_num_grad = FALSE,
  silent = TRUE,
  gamma_init = NULL,
  adaptive_weights = NULL,
  cap_loss = TRUE,
  gamma_init_strat = "smallest",
  developer_code = 0
)
}
\description{
Internal solver for fixed lambda
}
\keyword{internal}
