% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{krtexas_predict}
\alias{krtexas_predict}
\title{Predict from a KR TEXAS model}
\usage{
krtexas_predict(krtexas_model, newx)
}
\arguments{
\item{krtexas_model}{Model object returned by \code{krtexas_fit()}.}

\item{newx}{New data matrix with same number of columns as training X.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict from a KR TEXAS model
}
