% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{krtexas_predict_internal}
\alias{krtexas_predict_internal}
\title{Internal prediction function used by CV / fixed-lambda}
\usage{
krtexas_predict_internal(gammas_test, X_test, X_train, Y, A, kernel, sigma)
}
\description{
Internal prediction function used by CV / fixed-lambda
}
\keyword{internal}
