% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{loss_fun}
\alias{loss_fun}
\title{Gaussian kernel loss function (internal)}
\usage{
loss_fun(u, w, Ax, X, Y, sigma, kappa, alpha, adaptive_weights)
}
\description{
Gaussian kernel loss function (internal)
}
\keyword{(internal,}
\keyword{C++}
\keyword{Gaussian}
\keyword{function}
\keyword{internal}
\keyword{kernel}
\keyword{loss}
\keyword{wrapper)}
