% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{loss_fun_ep}
\alias{loss_fun_ep}
\title{Epanechnikov kernel loss (internal)}
\usage{
loss_fun_ep(u, w, Ax, X, Y, sigma, kappa, alpha)
}
\description{
Epanechnikov kernel loss (internal)
}
\keyword{internal}
