% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krtexas_functions.R
\name{loss_grad_fun}
\alias{loss_grad_fun}
\title{Gaussian kernel loss and gradient function together (internal)}
\usage{
loss_grad_fun(u, w, Ax, X, Y, sigma, kappa, alpha, adaptive_weights)
}
\description{
Gaussian kernel loss and gradient function together (internal)
}
\keyword{(internal,}
\keyword{C++}
\keyword{Gaussian}
\keyword{and}
\keyword{function}
\keyword{gradient}
\keyword{internal}
\keyword{kernel}
\keyword{loss}
\keyword{wrapper)}
